Redo tests for multiple dependencies
Surprisingly, a live explicit test for this does not exist in the test
suite.

There was a test that only used the local executor, and which was
disabled in 2018 in commit d7b865b309d3eac3eb6b2df6292fd8553f97af7c.
This PR removes that test file in favour of the new test.