Tidy DataFuture initialisation
This performs stronger runtime type checking on the initializer,
removes code paths where a DataFuture is constructed with a string
(which was turned into an error case in PR #1412)
and removes NotFutureError, as that type error is now handled by
the typeguard.typechecked decorator.

This comes from work on the benc-mypy branch.