Enable DFK type checking in dfk.__init__()
Adding type annotations turns on type checking for __init__() which then
means that many attributes of DFK objects now receive a type more specific
than Any, which can be checked through the rest of the codebase (with no
further changes for code which is already type-checked).

Because of that additional typechecking, a bunch of PRs earlier than this one
made several changes that were needed to allow type checking to actuall pass
when this PR is merged.

This PR also removes the piece of code which checks for dict based configs,
which was to help in the transition to class based configs, issue #133
in 2018. This code path is now unreachable due to typeguard issuing a
more generic error. Four and a half years is probably long enough to expect
users to no longer need this forward-migration assisting error.