Clarify that args can be any Sequence
The type of args is ambiguous within the DFK - someties its a list,
sometimes its a tuple...

Some parts of the DFK (for example TaskRecord) aready accomodate this
by using a Sequence.

This PR extends that to more places.

This is part of work towards a statically typed DFK: issue #2540