Rework checkpoint timer configuration for typing and usability
This PR changes error behaviour so that if the time string cannot be parsed,
an exception will be raised at DFK initialisation. The previous behaviour
on parse error was to log an error and use a default value instead.

This should be slightly improved usability by making misconfigurations more
obvious.

This is part of work towards a statically typed DFK: issue #2540