Add assertions for worker count manual test
Without these, the manual worker count test does not type check, as it
pulls an executor out of the DFK and makes an assumption that it is an
htex executor.

This assumption is not necessarily true: the DFK can have other executors
in it such as the local thread _parsl_internal executor.

These asserts clarify to the type check that these assumptions have been
made.