Add type annotations to Channels plugin superclass
There is an ambiguity here: walltime is described as an int, but has a
default value of None, which is not an int.

This default value isn't ever used, because execute_wait is always
overridden with a new type signature giving a new default, in subclasses.

So, should execute_wait walltime be typed as Optional[int] or int?

Option A:
With type Optional[int], all subclasses must now behave correctly when
given a None walltime.

The existing subclasses are implemented expecting an int, with their own
default values.

Option B:
With type int, the default None value needs to be replaced with an int.
This int is never used in practice because the value will be overridden in
subclasses.

So, this PR changes the default None value to 0, and assumes that a
different default will be specified in Channel subclasses, as happens
already.

A similar ambiguity exists with makedirs.mode