Rework _work_queue_submit_wait arguments
This PR makes all arguments into mandatory keyword arguments,
and adds type annotations.

The default values are removed because they're always specified
at the call site, and some defaults are dangerous: it is not valid
to share a multiprocessing.Queue between multiple instances of the
submit/wait thread.

The env parameter appears to be never specified, and so this PR removes that
parameter entirely.

To typecheck task.env_pkg usage, the package_run_script parameter needs to
not be None when that is used, and so this PR extends the existing task
submission failure error handling to include a new exception that can be
raised there.

This was driven by work in the benc-mypy branch