Remove unused incompatible run_dir definition from WQ Executor
In practice, at runtime this incompatible definition was overridden during
DFK initializtion, to be replaced by a string (not a function) that points
to the rundir.

So the code being remove by this PR, at runtime, is in effect dead code that
never gets invoked.

However, for typechecking purposes, this leads to incorrect assumptions about
the type of run_dir - that it has the type of the removed function, rather
than being a string property (as should be inherited from the ParslExecutor
superclass)

This PR is part of work to introduce static typing to the Work Queue
executor.