Remove manager to interchange work requests to speedup sequential task distribution (#2588)
Prior to this PR, the interchange calculated how many tasks to send to a manager based on the minimum of: how many tasks the manager most recently requested (minus sent tasks); how many tasks the manager has total capacity for, minus the number of tasks in flight.

The most recently requested count is made on an exponential backoff, which continues to backoff when a manager is saturated with long tasks. That means that the requested task count is sometimes not updated for quite a long time.

This PR removes that clause entirely, and the code that delivers that information. Now the interchange relies only on the computed maximum capacity minus the number of tasks in flight - this code was added in PR #881, but alongside the previous "manager requests work" value.

A new task will be dispatched when a result arrives from a manager, freeing up a slot, rather than on a separate free_capacity poll: so new task dispatch is now much more event driven.

On my laptop on master prior to this PR, some test code shows that when tasks execute quickly, tasks begin executing just over 1 or just over 2 poll_periods after submission; but when tasks take a long time and a manager is saturated, backoff of poll periods in the manager means that the manager remains idle for sometimes 9 seconds before starting to execute a task.

After this PR, the same code shows tasks executing in just over one poll period in both situations.
I have tried this with a range of poll periods from 10 to 500ms.

With smaller poll periods, 1 .. 3 ms, I have noticed that the slow task saturated case takes a few ms longer for task dispatch compared to the fast task path. I have not investigated why that happens, and I don't intend to for this PR.

@WardLT noticed this behaviour in May 2022, and it is still present at time of writing this PR.