Fix modification of ignore_for_cache when using AUTO_LOGNAME
When using AUTO_LOGNAME, stderr/stdout are added to ignore_for_cache,
because this parameter is different on each invocation, and so
invocations will not be cached.

However, what is modified is the ignore_for_cache structure in the
function itself, not a local variable inside the invocation, which
means that it is modified for all subsequent calls, including calls
which do not use AUTO_LOGNAME.

This was found by mypy on the benc-mypy branch.