Remove join from python_app decorator, and reimplement join_app
The join parameter should never be set by a user, and gives bad error
messages if it is set - see issue #2147.

The parameter was only there so that the join_app decorator could be
implemented in terms of the python_app decorator. This PR changes how
that implementation works so that this misbehaviour is not exposed
to end users.

Fixes issue #2147