Make harder attempt to find the source file for a documentation link
Prior to this, a URL pointing into github was generated directly from the
name of a module.

Sometimes this pointed to a file that re-exported the symbol, not the
original source - for example, referring to parsl.providers.CondorProvider
when the source lives in parsl.providers.condor.condor

Sometimes that file did not exist at all - for example, referring to
parsl/providers.py when that module is defined in parsl/providers/__init__.py

This PR attempts to bundle together a few heuristics to make source file
location work better.

See issue #2604