Make Work Queue executor source code more explicit about task types
There are three different entities with the word "task" used to
describe them: parsl tasks, tasks that go in this executor's
self.tasks dictionary, and Work Queue tasks.

The executor code refers to two of them, and this PR attempts to be explicit
in every log line, and in some variable names, about which kind of task is
being referred to.

self.tasks dictionary tasks are referred to as "executor tasks".

Work Queue tasks are referred to as Work Queue tasks.

This is part of an attempt to make log files easier to understand in the
presence of many kinds of task.