Explictly render States.__str__ like in pre-python 3.11
Python 3.11 changes how IntEnum, the parent class of States, shows states
using __str__, so that only the integer value is returned. This change was
made to support a usecase that is not the parsl usecase.

This new __str__ rendering is less user friendly, as users usually
probably prefer to see textual state names.

This change makes the pre-python 3.11 behaviour happen on all
Python versions, so that states continue to appear as

States.exec_done

rather than as an integer.