Support python 3.11 in CI (#2461)
This PR makes a change to the task state enum so that it renders the same across all Python versions, the same as pre-3.11. Python 3.11 changes how enumerations are rendered by default by __str__ to support a usecase that is not relevant here (to support porting from int constants to enums).