Improved extraction of the job ID from sbatch stdout (#2636)
On the HPC I'm working (VSC clusters in Flemish universities, Belgium), the output of the sbatch command is slightly different from the default and therefore not processed correctly. The consequence is that submitted jobs for workers are not terminated when Parsl shuts down. This error did not result in an error message in parsl.log.

This is an example of the output of sbatch on one of the VSC clusters:

Submitted batch job 60060368 on cluster slaking
The fix proposed in this PR introduces a minimal customization with a more robust default than the current implementation. The proposed default is consistent with the current behavior and works for us too.