Make htex.connected_managers IPC call into a function
Prior to this PR, htex.connected_managers was a @property, which makes
reasoning about the effect behaviour of code using this property harder
to reason about:

For example, in htex scaling down code, when the interchange code is gone
away, the value of executor.connected_managers is a silent hang, not a value.

Two other IPC @property values are not changed by this PR, connected_workers
and outstanding, because they are exposed outside of the executor, to
core scaling code, and I would like to think about that a bit more.