Make interchange threads daemons for less-hangy exit (#2637)
When the interchange is killed by SIGINT/Ctrl-C, the main thread exits (due to a SIGINT generated KeyboardInterrupt exception) but the other two interchange threads do not. This means that the interchange process hangs without exiting.

With the behaviour introduced by this PR, the main thread ending will now cause the (now daemon) threads to end too.

This behaviour is different from the SIGTERM behaviour (which is how a clean parsl shutdown happens) - a SIGTERM kills all threads without opportunity for catching the exit, in the current signal handler configuration.

To recreate:

run:

pytest parsl/tests/ --config parsl/tests/configs/htex_local_alternate.py  -k 'not cleannet'
press ctrl-C after a while

observe a hang and that the interchange process is still alive.

After this PR,

observe a hang and that the interchange process has terminated.

The pytest hang in this post-PR situation comes from a couple of other places at least, that this PR does not address, but it's no longer in the interchange.