Traverse configuration heirarchy to report more usage information
This PR introduces code to traverse the configuration object (in a similar
manner to the RepresentationMixin style of logging the supplied configuration
object) with the intention of giving each object a chance to report its
own usage information.

This PR modifies the HighThroughputExecutor to use this API to report
richer usage information: a specific usage query is to ask about use of the
enable_mpi_mode parameter and this modification supports that.

Beware that this reports on configuration of components, and does not
report any further usage unless those components are so augmented using
the new API: for example, configurations by default will include three
staging providers, even though I believe it is extremely rare that either
the FTP or HTTP staging providers are actually used to stage data.
(It is hopefully a straightforward change to add a UsageInformation
implementation to report if those classes are actually used to stage
data in any run).

To support UsageInformation instances which report on usage during a run,
the component tree is traversed both for the start message and the end
message, and may result in different information in each message.

An example start message looks like this: (pretty-formatted)

{'correlator': 'f7595d08-7b94-49bc-b3d7-1ea7532b2f51',
 'parsl_v': '1.3.0-dev',
 'python_v': '3.12.2',
 'platform.system': 'Linux',
 'start': 1710150467,
 'components': ['parsl.config.Config',
                {'c': 'parsl.executors.high_throughput.executor.HighThroughputExecutor', 'mpi': False},
                'parsl.providers.local.local.LocalProvider',
                'parsl.channels.local.local.LocalChannel',
                'parsl.launchers.launchers.SingleNodeLauncher',
                'parsl.data_provider.ftp.FTPInTaskStaging',
                'parsl.data_provider.http.HTTPInTaskStaging',
                'parsl.data_provider.file_noop.NoOpFileStaging',
                'parsl.monitoring.monitoring.MonitoringHub']}