Allow join_apps to return multiple Futures (#2167)
To make a join_app wait for multiple component apps, return a list of Futures from the join app, instead of a single Future.

Previously, to wait for multiple launched Futures inside an app, it was necessary to make a separate combine app which waits for them all the complete.

Use of this style of additional trivial app caused awkwardness in monitoring and checkpointing: it often did not have a clear separate identity based on parameters, which made it hard to distinguish "different" combine calls from each other in the monitoring and checkpointing databases. This was also boilerplate that seems nicer to not force on the user.

This PR also moves sending monitoring updates more strictly before adding callbacks: in the case of fast executing tasks, a race condition would mean that tasks would complete (and callbacks fire) before handling of the join app completion code finished. This resulted in two exec_done messages, rather than one joining and one exec_done message.