Implement JobState enum like dataflow States (#2653)
This PR changes the implementation of the JobState enumeration to be implemented the same way (using an IntEnum) as the parsl.dataflow.states.States enumeration for task states.

The previous implementation added an extra field in order to tag whether an enumeration value is a final value or not. states.States implements this by specifying a list of terminal states, distinct from the state values, and this PR makes JobState be implemented this way.

The previous implementation broke (see issue #2652) due to enumeration implementation changes in Python 3.11. This newer implementation is, hopefully, easier to understand.

This PR also adds a regression test for #2652 to check that repr does not raise an exception, and additionally checks that the string representation of JobState is consistent across Python versions.