Remove Python 2 (object) explicit superclasses for consistency
These are used in Python 2 to indicate that "new style" classes should be
used. Python 3 does not have a new/old style class distinction and object
is the default superclass of all classes.

There is no consistent approach in the parsl codebase to indicating
an object superclass explicitly or not, and this PR attempts to make the
code more consistently not do so.

This should not affect code behaviour.