Remove unused htex executor.is_alive attribute (#2674)
This has never been used in htex since it was introduced, as far as
I can tell. It looks like it was once part of an attempt to shut down
the queue management thread, but it is not used that way and the
queue management thread is always left running until the end of the
main python process.

This PR removes any pretence that thread shutdown happens. A future
PR might introduce a working thread shutdown.