Fix regression test for issue #97 - parallelism=0
Previously, The test was setting incorrect attributes (on an executor, not
a provider), and was using an executor which did not use scaling (the local
executor).

This PR switches to using the high throughput executor, and configures it
directly, rather than modifying an existing configuration.

This PR simplifies the activity performed in the test, to focus on what
is being tested: that a single task can run.

This PR was driven by an upcoming PR which increases mypy coverage in the
test suite, and which revealed type errors even though the test was
disabled.

I checked that this does detect the regression in questions by disabling
scaling case 4a in parsl/dataflow/strategy.py - which was the code block
added to address issue #97.