Add test that visualization coloring tables color each known state
This is a regression test for PR #2708, issue #2681, that will also
detect other new states added to the states enumeration without being
added to the coloring tables.

In order to be able to test this way, this PR moves the two coloring tables
to be module scope, instead of local scope inside a method.