Allow executors specification to be a single string
This PR also prohibits executors from being labelled 'all', which is
backwards incompatible (although probably minor). This prohibition is
because the specification string 'all' for all executors now overlaps
the executor label namespace, and so ambiguity would be possible.

This PR is intended to address minor syntactic noise in the common case
of a single executor being specified:

Before:
@python_app(executors=['SuperComputer1'])

After:
@python_app(executors='SuperComputer1')

This PR also adds test cases around the syntax for executors.