Remove DFK parameter from strategy to facilitate use without DFK (#2712)
For example, the funcX project to use scaling with their use of the
HighThroughputExecutor without a DFK.

This PR makes a behavioural change: executors from the DFK are no
longer added in the strategy initialiser. The DFK already adds a
executors via JobStatusPoller: prior to this PR, this meant executors
were added to the strategy twice, but after this PR, this means that
executors are only added once - by the job status poller.