Check that serializers register with an identifier of length 3 (#2716)
It is required in the code that all serializers register with an identifier
of length 3 bytes, but this is not checked, and instead the code assumes that
the length of the identifier of the "first" serializer is the length of all
identifiers. If this is not true, then the code misbehaves mysteriously.

This assert helps future programmers discover that implicit requirement.