Force two iterations in parsl-perf for proper warmup
Prior to this PR, if the first iteration takes longer than the target
time (120s by default) due to warmup, rather than due to raw task
throughput, the task rate from the first iteration will be used.

This PR forces two iterations, so that the first iteration forces warmup
and a second iteration always to make a meaurement in warmed-up state.