Rewrite file paths in WorkQueue coprocess mode, for non shared-fs
Prior to this, Work Queue staged the relevant function files in (and
attempted to stage the missing result files out), but the coprocess
worker code was given the original (non-staged) paths.

This worked in shared-fs environments: the staged files were ignored,
and the original files were used, but broke in non-shared-fs environments.

This PR uses the staged filename (the basename), as already happens with
non-coprocess mode.

Fixes #2728