Make resource monitoring process a daemon
Prior to this PR, if the parent task executing process exits without
telling the resource monitoring process to exit, that parent process
will hang at exit waiting for the resource monitoring process to exit,
which will never happen - because the resource monitoring process is
expecting the parent process to tell it when to exit.

In daemon mode, the resource monitoring process will be terminated and
the parent process will not hang in this way.

When exiting in this way, the resource monitoring process will not
send final resource usage information records, rather than continuing
to send that information forever. That may lose track of excessive
sudden resource usage that might be useful in debugging the unexpected
parent process exit; but is a tradeoff against hanging process exit.

In normal operation, this change should not affect anything else:
the shutdown procedure both before and after this PR is that the parent
process sends an event to the resource monitoring process to exit,
waits up to 30 seconds for it to exit and forcibly kills it if it has
not exited.

The main problem with this is a misconfigured environment resuling in
an unanticipated exception, documented in issue #2754