Permit both typeguard 2.x and 4.x
This PR loosens some type checking to accomodate both typeguard 2.x and 4.x
as large parsl users (eg. LSST DESC) still package version 2.x but other users
want the latest - 4.x

In typeguard 4.x, local variables are instrumented for type-checking, which
is incompatible with using multiprocessing.Queue: multiprocessing.Queue is
not a type as far as typeguard is concerned, so a single use of that, in
MonitoringHub, is commented out by this PR.

Typeguard 2.x and 4.x have different type error classes, so explicit type
checking of exception classes cannot happen in tests as the typeguard 4.x
TypeCheckError class does not exist when typeguard 2.x is installed. This
PR makes that type-check of type-checking errors happen using string
matching of the name of the type-check error class in
test_bash_apps/test_stdout.py

See for example issue #3017

This PR is ugly but attempts to be pragmatic about user packaging needs.