Update python_apps/test_type5
In another test (`test_futures`), we already test the delay (apps finish in
non-deterministic order), so no sense in arbitrarily sleeping during these
tests.  Meanwhile, the thrust of these tests appears to be in proving that the
_linkage_ between functions is automatic, given futures as app/function
arguments.  So, faithfully implement that in a more Pythonic way, and
parameterize for some variability.

Faster, no extraneous prints, and no pytest warnings.