Remove bad monitoring type annotation (#2762)
This annotation was added in the assumption that ForkProcess is a descendent
of Process. This is not the case for a couple of reasons:

* ForkProcess is a type that is dynamically computed at runtime (by asking the
multiprocess module for a relevant type, in parsl.multiprocessing).

* The types that are actually/usually returned are descendents of the
non-documented multiprocessing.BaseProcess class, not of
multiprocessing.Process.

Because ForkProcess is dynamically constructed, it cannot be used as the type
annotation for the p variable.

mypy does not strictly typecheck ForkProcess, so this incoherence was not a
problem, but more recent (>=3) versions of typeguard discover this incoherence
at runtime and raise a runtime type error.

Crossref issue #2754 where a user accidentally installing a newer version of
typeguard caused the relevant typechecking error to occur.

Crossref issue #2343 for other fork related issues - not directly this problem
but perhaps useful to bear in mind.

This PR removes the bad type annotation entirely.