Rearrange serializer caching to fix issue #2555
Prior to this PR:
parsl.serialize.facade monkeypatches some serializer objects to create
new serializers, but does not generate new identifiers for those new
serializers, so two serializers exist in the serializer globals with
the same identifier.

This patching is done to make some new serializers that have caching
behaviour.

This leads to issue #2555, where data from one serializer is deserialized
using one of the these dynamically defined serializers, incorrectly.

This PR removes that monkeypatching and moves caching behaviour into
two new serializers explicitly defined for the type of objects that should
be cached.

This adds two new serializer identifiers, which means there is an
incompatible wire protocol change: older versions of parsl will not be able
to deserialize callables serialised after this PR.

This PR adds a regression test for issue #2555 that fails before this PR.

Fixes issue #2555