serializer plugin API
TODO: (broken) make a test that runs using a custom serializer to an htex worker... to check that remote processes can load a custom serializer.

prior to this PR, serializer registration happened in order of class definiton, which in the case of a single concretes file, is the order of the serializers inside that file.

this does not work well when trying to define other serializers in other files, because the order of import of files can change subtly and because of distant effects.

so doing serialiser registration in a subclass hook is not a very nice thing to do here...

this PR should move to a model which defines the default de*serializers explicitly, and allows the user to modify that list explicitly too, to add in new de*serializers.

limitations:

this is part of work needed to allow serializer plugins, but not all of it, as it does not define how a remote worker will discover plugged-in serializers - for that, perhaps, the header should become something importable? instead of a registry-based ID bytes string.
contrasting with that: proxystore doesn't need its own deserializer! it can use the existing pickle/picklecode deserialisers, because what it generates is a pickled object... and that's a legitimate use case... which suggests that serializers don't need to be indexed by "serializer" ID at all on the serializing side, but only on the deserializing side...

this API also will not easily allow serializer methods to be defined for values returned from workers to the submit side: the expected place of registration, near the start of the user workflow, does not have a corresponding place on the worker side.

this PR separates out serialization behaviour (object -> bytestream) from deserialization
behaviour: (identifier, bytestream) -> object, because serializer might not (and in this
prototype, *cannot*) implement custom deserialization... due to non-registration on the
remote side, and instead can only generate dill/pickle targeted bytestreams


the motivating use case for this basic serializer plugin API is supporting proxystore, prototyped in
https://github.com/Parsl/parsl/pull/2718
https://labs.globus.org/projects/proxystore.html

this PR also contain an optional proxystore plugin that would need to be activated by a user,
and an attempt at an in-pickle (recursive style) proxystore plugin that is policy-aware and can
proxy stuff deeper inside lists... to deal with proxying only interesting objects deep inside
graphs - which is actually what happens with regular argument lists when they're non-trivial.

loading remote serialisers via importlib is a lot of stuff... perhaps i should... just pickle the deserializer and send it (and let pickle deal with that?)
- i guess there's two paths here: one is loading remote deserialisers, and the other is focusing on the proxystore use case, which, differently, wants access to the pickle deserializer remotely - and so that drives clearer separation of serializers vs deserialisers... that 2nd case is probably waht i should concentrate on...

right now this does module loading in order to work with the class based serializer design inherited from before... but maybe that's not how it should end up...

perhaps just serializer callables and deserializer callables? or based around modules?

it seems nice to be able to configure the serializer (aka make a class or a partially applied callable) for configuring proxystore... but unclear how that would work for remote end configuration.