Add a readthedocs config, to fix bad python version and because it will soon be mandatory (#2805)
Prior to this PR, readthedocs was using Python 3.7 by default to build the docs. This became unsupported in parsl in PR #2749 and broken in Parsl in PR #2744.

This PR adds a configuration file to use Python 3.11, the latest supported version of Parsl, to build on readthedocs.

This configuration file will become mandatory in readthedocs in September 2023, according to their own documentation, so this file would need adding soon anyway.