Reduce test_scale_out delay
Locally, I was consistently seeing this test take 56s.  This commit reduces
that to 12s locally.

There are two main changes:

- Bind to 127.0.0.1, rather than to all live interfaces.  Some of the delay was
  from the network stack

- Replace the littered `time.sleep()` calls with an "Event" based approach
  (polled files on the filesystem)

The other addition of note is `try_assert()`.  This my approximation of JUnit's
`assertTimeout` or QT's `QTRY_VERIFY`: continually try a test condition for a
timeout period.

I'm not able to discern at this time if this robustly resolves #1885.