Reduce test delay from 45s to 1
The key change is to abuse the `JobStatusPoller.poll()` to ping the scale-out
logic more often than the hard-coded 5s.  Ping it every tenth of a second for
the second it takes the app() to complete.  Nominally, if the scaling logic is
correct, then with a single app it won't matter that it's spammed: we'll only
scale out a single block total.