Remove pickle serializer for callables
This serializer, in current practical parsl cases, will raise an
exception every time and fall through to the dill callable serializer,
because it cannot serialize the wrappers that Parsl user functions
are wrapped in.

Were it actually able to serialize, it would in many usecases generate
invalid serializations, rather than raising an exception: in any case
where an app was defined in a main file rather than in an importable
file (eg. simple workflow scripts, REPLs like Jupyter notebooks)