Make serializer registration explicit (#2831)
Previously registration was implicit in the class definition, using a
subclass hook. This makes it a bit awkward to work with the set of
registered serializers:

* serializers would be registered at definition time (which because of
non-deterministic import order) makes the order awkward to understand
if serialisers are defined in more than one module.

* serializers could not be defined without being registered

This PR makes the set of serializers more explicit, with the defaults
registered at import of the facade module. Definitions of serializers in
concretes.py (or anywhere else) now do not result in a registration.

This is part of a move towards user-pluggable serializers, although this
PR does not actually support adding more user serializers (as they will
not be imported on the distant end).