Increase log level for binding between WQ and executor ID namespaces
This is to help understand the connection between Work Queue and executor
level tasks when looking through logs.

Awkardly this binding is only logged at the end of a task execution, but
this is intended to be a minimal PR.

See for example #2313 for similar binding work in the DataFlowKernel.