Allow user pluggable serializers and add a demonstration ProxyStore serializer (#2842)
This PR rearranges the deserializer identifier handling so as to dynamically load unknown serializers. These should be identified by a module and class name (and that identifier will be generated automatically).

New serializers, implementing SerializerBase, can be implemented and registered for either the code or data paths in serialization using the existing register_method_for_* functions introduced in #2831

This PR also introduces an example serializer using this API: it extends dill to use ProxyStore on certain objects determined by a user specified policy.

The API for reconfiguring serializers is rather awkward, as evidenced by one of the ProxyStore test case, but this PR does not attempt to make this any nicer.