Reduce monitoring filesystem radio log volume
This PR moves most log messages in the filesystem radio to debug level,
and sets the log level for the monitoring to INFO level. This gives one
log message per monitoring message, without per-poll logging. This is
more in line with log level styles in Parsl, logging at INFO when something
happens.

Output of DEBUG logs is not user-configurable, but I left those log lines
there for debug hacking.

Fixes #2845